// 2026-07-12_tiki_tiki/src/types.ts
export interface Video {
    id: string
    username: string
    displayName: string
    avatarUrl: string
    caption: string
    musicId: string
    musicName: string
    videoUrl: string
    isFollowed: boolean
    likeAmount: number
    isLiked: boolean
    commentAmount: number
    bookmarksAmount: number
    isBookmarked: boolean
    shareAmount: number

}